separate navigation bar as its own component
