+++ client/src/app/components/footer/footer.component.html
@@ -1,9 +1,10 @@
 <footer class="flex items-center justify-center h-6 bg-gray-100 border-t border-gray-200 text-xs text-gray-500">
   <div class="flex items-center gap-2">
+    <a routerLink="/about" class="hover:text-gray-700 hover:underline">About</a>
+    <span class="text-gray-300">|</span>
     <a routerLink="/impressum" class="hover:text-gray-700 hover:underline">Impressum</a>
     <span class="text-gray-300">|</span>
     <a routerLink="/privacy" class="hover:text-gray-700 hover:underline">Privacy Policy</a>
     <span class="text-gray-300">|</span>
-    <a routerLink="/about" class="hover:text-gray-700 hover:underline">About</a>
   </div>
 </footer>
