+++ client/src/app/components/navigation-bar/navigation-bar.component.html
@@ -1,0 +1,32 @@
+<div class="flex flex-col h-full rounded-2xl p-5 mr-3 bg-gray-100 items-center justify-between">
+  <!-- Helios Icon -->
+  <app-helios-icon routerLink="/" size="3rem" class="rounded-xl w-12 cursor-pointer" />
+  <p-divider />
+  @if (repositoryId()) {
+    <!-- Repository Initials -->
+    <p-avatar label="{{ repositoryQuery.data()?.name | slice: 0 : 1 }}" pTooltip="{{ repositoryQuery.data()?.nameWithOwner }}" size="large" />
+    <p-divider />
+
+    <!-- Repository pages -->
+    <div class="flex flex-col gap-3">
+      @for (item of items(); track item.label) {
+        <a
+          class="rounded-xl p-2 flex items-center text-gray-500 hover:text-gray-700"
+          [routerLink]="item.path"
+          [routerLinkActive]="'!bg-slate-800 !text-slate-100 !hover:text-slate-100'"
+          [pTooltip]="item.label"
+        >
+          <i-tabler [name]="item.icon" class="!size-10 !stroke-1" />
+        </a>
+      }
+    </div>
+  }
+  <span class="flex-grow"></span>
+  @if (repositoryId()) {
+    <!-- Lock Info -->
+    <app-user-lock-info />
+  }
+  <p-divider />
+  <!-- Login/Logout Button -->
+  <app-profile-nav-section />
+</div>
