+++ client/src/app/components/profile-nav-section/profile-nav-section.component.html
@@ -1,4 +1,3 @@
-<p-divider />
 <div class="flex flex-col gap-3 items-center">
   @if (isLoggedIn()) {
     <p-avatar shape="circle" size="normal" [image]="getProfilePictureUrl()" [pTooltip]="fullName()" tooltipPosition="top" (click)="openProfile()"> </p-avatar>
