+++ client/src/app/pages/main-layout/main-layout.component.html
@@ -8,26 +8,8 @@
   }
   <div class="flex flex-grow p-3" [ngClass]="isLoggedIn() ? 'h-[calc(100vh-24px)]' : 'h-[calc(100vh-48px)]'">
     <!-- Fixed Navigation Bar -->
-    <div class="rounded-2xl p-5 mr-3 bg-gray-100 flex flex-col items-center justify-between h-full sticky top-3">
-      <app-helios-icon routerLink="/" size="3rem" class="rounded-xl w-12 cursor-pointer" />
-      <p-divider />
-      <p-avatar label="{{ repositoryQuery.data()?.name | slice: 0 : 1 }}" pTooltip="{{ repositoryQuery.data()?.nameWithOwner }}" size="large" />
-      <p-divider />
-      <div class="flex flex-col gap-3">
-        @for (item of items(); track item.label) {
-          <a
-            class="rounded-xl p-2 flex items-center text-gray-500 hover:text-gray-700"
-            [routerLink]="item.path"
-            [routerLinkActive]="'!bg-slate-800 !text-slate-100 !hover:text-slate-100'"
-            [pTooltip]="item.label"
-          >
-            <i-tabler [name]="item.icon" class="!size-10 !stroke-1" />
-          </a>
-        }
-      </div>
-      <span class="flex-grow"></span>
-      <app-user-lock-info />
-      <app-profile-nav-section />
+    <div class="h-full sticky top-3">
+      <app-navigation-bar [repositoryId]="repositoryId()"></app-navigation-bar>
     </div>
 
     <!-- Scrollable Content Area -->
